feat: Includes first working version of algorithm
This commit includes the first working version of the Particle Filter.

It follows a naive approach of brute-force searching for the nearest landmark
given a set of observations and particles. This approach is sufficient for the
reduced dimension problem we have, but would probably have performance issues
in a higher dimensional space.